fix typo in package name
